+++ android/app/src/main/java/app/priceguard/di/ProductRepositoryModule.kt
@@ -2,9 +2,9 @@ package app.priceguard.di
 
 import app.priceguard.data.GraphDataConverter
 import app.priceguard.data.network.ProductAPI
-import app.priceguard.data.repository.ProductRepository
-import app.priceguard.data.repository.ProductRepositoryImpl
-import app.priceguard.data.repository.TokenRepository
+import app.priceguard.data.repository.product.ProductRepository
+import app.priceguard.data.repository.product.ProductRepositoryImpl
+import app.priceguard.data.repository.token.TokenRepository
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
