+++ android/app/src/main/java/app/priceguard/ui/home/list/ProductListFragment.kt
@@ -9,8 +9,8 @@ import androidx.coordinatorlayout.widget.CoordinatorLayout
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import app.priceguard.R
-import app.priceguard.data.repository.ProductErrorState
-import app.priceguard.data.repository.TokenRepository
+import app.priceguard.data.repository.product.ProductErrorState
+import app.priceguard.data.repository.token.TokenRepository
 import app.priceguard.databinding.FragmentProductListBinding
 import app.priceguard.ui.additem.AddItemActivity
 import app.priceguard.ui.detail.DetailActivity
